Remove item ability added
I've introduced a new `rm` flag to the function to be able to remove items from localStorage, reactively. This means, of course, that people will have to pass in an arbitrary `val` to remove items, but I couldn't find a better way without breaking the API as it is available. I can attach a function to the prototype of this class, but wanted to know you guys' opinion.